Extract debugging a parameter based classes to separate files
