+++ lib/ConfigParameter/BaseParameter.h
@@ -1,0 +1,18 @@
+#ifndef __BASE_PARAMETER_H__
+#define __BASE_PARAMETER_H__
+
+enum Mode {ap, api};
+enum ParameterMode { get, set, both};
+
+/**
+ * Base Parameter
+ */
+class BaseParameter {
+public:
+    virtual ParameterMode getMode() = 0;
+    virtual void fromJson(JsonObject *json) = 0;
+    virtual void toJson(JsonObject *json) = 0;
+    virtual void clearData() = 0;
+};
+
+#endif /* __BASE_PARAMETER_H__ */
