+++ src/ConfigManager.h
@@ -24,6 +24,10 @@
     #define WIFI_OPEN  WIFI_AUTH_OPEN
 #endif
 
+#include "DebugPrint.h"
+#include "ConfigParameter.h"
+#include "ConfigStringParameter.h"
+
 #define MAGIC_LENGTH 2
 #define SSID_LENGTH 32
 #define PASSWORD_LENGTH 64
@@ -33,112 +37,11 @@
     using WebServer = ESP8266WebServer;
 #endif
 
-extern bool DEBUG_MODE;
-
-#define DebugPrintln(a) (DEBUG_MODE ? Serial.println(a) : false)
-#define DebugPrint(a) (DEBUG_MODE ? Serial.print(a) : false)
-
 extern const char mimeHTML[];
 extern const char mimeJSON[];
 extern const char mimePlain[];
 extern const char mimeCSS[];
 extern const char mimeJS[];
-
-enum Mode {ap, api};
-enum ParameterMode { get, set, both};
-
-/**
- * Base Parameter
- */
-class BaseParameter {
-public:
-    virtual ParameterMode getMode() = 0;
-    virtual void fromJson(JsonObject *json) = 0;
-    virtual void toJson(JsonObject *json) = 0;
-    virtual void clearData() = 0;
-};
-
-/**
- * Config Parameter
- */
-template<typename T>
-class ConfigParameter : public BaseParameter {
-public:
-    ConfigParameter(const char *name, T *ptr, ParameterMode mode = both) {
-        this->name = name;
-        this->ptr = ptr;
-        this->mode = mode;
-    }
-
-    ParameterMode getMode() {
-        return this->mode;
-    }
-
-    void fromJson(JsonObject *json) {
-        if (json->containsKey(name) && json->is<T>(name)) {
-            *ptr = json->get<T>(name);
-        }
-    }
-
-    void toJson(JsonObject *json) {
-        json->set(name, *ptr);
-    }
-
-    void clearData() {
-        DebugPrint("Clearing: ");
-        DebugPrintln(name);
-        *ptr = T();
-    }
-
-private:
-    const char *name;
-    T *ptr;
-    std::function<void(const char*)> cb;
-    ParameterMode mode;
-};
-
-/**
- * Config String Parameter
- */
-class ConfigStringParameter : public BaseParameter {
-public:
-    ConfigStringParameter(const char *name, char *ptr, size_t length, ParameterMode mode = both) {
-        this->name = name;
-        this->ptr = ptr;
-        this->length = length;
-        this->mode = mode;
-    }
-
-    ParameterMode getMode() {
-        return this->mode;
-    }
-
-    void fromJson(JsonObject *json) {
-        if (json->containsKey(name) && json->is<char *>(name)) {
-            const char * value = json->get<const char *>(name);
-
-            memset(ptr, 0, length);
-            strncpy(ptr, const_cast<char*>(value), length - 1);
-        }
-    }
-
-    void toJson(JsonObject *json) {
-        json->set(name, ptr);
-    }
-
-    void clearData() {
-        DebugPrint("Clearing: ");
-        DebugPrintln(name);
-        memset(ptr, 0, length);
-    }
-
-
-private:
-    const char *name;
-    char *ptr;
-    size_t length;
-    ParameterMode mode;
-};
 
 /**
  * Config Manager
